--- app/src/main/java/com/otus/homework/chessclient/onboarding/presenters/ILoginPresenter.kt
@@ -1,7 +1,0 @@
-package com.otus.homework.chessclient.onboarding.presenters
-
-import com.otus.homework.chessclient.core.presenters.IPresenter
-
-interface ILoginPresenter : IPresenter{
-
-}
