--- app/src/main/java/com/otus/homework/chessclient/onboarding/presenters/IRegistrationPresenter.kt
@@ -1,6 +1,0 @@
-package com.otus.homework.chessclient.onboarding.presenters
-
-import com.otus.homework.chessclient.core.presenters.IPresenter
-
-interface IRegistrationPresenter : IPresenter {
-}
