+++ app/src/main/java/com/otus/homework/chessclient/onboarding/views/RegisterFragment.kt
@@ -1,20 +1,18 @@
 package com.otus.homework.chessclient.onboarding.views
 
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.fragment.findNavController
 import com.jakewharton.rxbinding2.view.RxView
 import com.jakewharton.rxbinding2.widget.RxTextView
 import com.otus.homework.chessclient.R
 import com.otus.homework.chessclient.onboarding.model.News
 import com.otus.homework.chessclient.onboarding.model.enums.NewsMessageId
-import com.otus.homework.chessclient.onboarding.presenters.IRegistrationPresenter
+import com.otus.homework.chessclient.onboarding.presenters.RegistrationPresenter
 import com.otus.homework.model.enums.AppScreens
 import io.reactivex.Observable
 import io.reactivex.functions.BiFunction
@@ -28,7 +26,7 @@ import java.util.concurrent.TimeUnit
 class RegisterFragment : Fragment(), IRegisterView, KodeinAware {
     override val kodein: Kodein by kodein()
 
-    private val presenter: IRegistrationPresenter by instance()
+    private val presenter: RegistrationPresenter by instance()
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         return inflater.inflate(R.layout.fragment_registration, container, false)
