+++ app/src/main/java/com/otus/homework/model/task/ChessFigure.kt
@@ -3,6 +3,8 @@ package com.otus.homework.model.task
 import com.otus.homework.model.enums.ChessFigureColor
 import com.otus.homework.model.enums.ChessFigureType
 
-data class ChessFigure (val figureType:ChessFigureType,
-                        val color:ChessFigureColor,
-                        val position: FigurePosition)
+data class ChessFigure (
+    val figureType:ChessFigureType,
+    val color:ChessFigureColor,
+    val position: FigurePosition
+)
