+++ app/src/main/java/com/otus/homework/model/task/ChessTask.kt
@@ -2,6 +2,8 @@ package com.otus.homework.model.task
 
 import com.otus.homework.model.enums.ChessFigureColor
 
-data class ChessTask(val id:String,
-                     val startingPositions:List<ChessFigure>,
-                     val activeColor:ChessFigureColor)
+data class ChessTask(
+    val id:String,
+    val startingPositions:List<ChessFigure>,
+    val activeColor:ChessFigureColor
+)
