+++ app/src/main/java/com/otus/homework/model/task/FigurePosition.kt
@@ -1,3 +1,6 @@
 package com.otus.homework.model.task
 
-data class FigurePosition (val row:Int = -1, val column:Int = -1)
+data class FigurePosition (
+    val row:Int = -1,
+    val column:Int = -1
+)
