+++ app/src/main/java/com/otus/homework/network/OnBoardingApi.kt
@@ -3,13 +3,13 @@ package com.otus.homework.network
 import com.otus.homework.model.enums.ChessTaskDifficulty
 import com.otus.homework.model.task.ChessTask
 import com.otus.homework.model.user.UserShortData
-import com.otus.homework.network.interfaces.IOnBoardingApi
-import com.otus.homework.network.interfaces.IRetrofitBuilder
+import com.otus.homework.network.interfaces.UserApi
+import com.otus.homework.network.interfaces.NetworkEngineBuilder
 import io.reactivex.Single
 import okhttp3.ResponseBody
 import retrofit2.Response
 
-class OnBoardingApi(private val builder:IRetrofitBuilder) : IOnBoardingApi {
+class OnBoardingApi(private val builder:NetworkEngineBuilder) : UserApi {
     private val service = builder.buildOnBoardingService()
 
     override fun login(newUserData: UserShortData): Single<Response<ResponseBody>> = service.loginUser(newUserData)
