+++ app/src/main/java/com/otus/homework/network/interfaces/NetworkEngineBuilder.kt
@@ -1,5 +1,5 @@
 package com.otus.homework.network.interfaces
 
-interface IRetrofitBuilder {
+interface NetworkEngineBuilder {
     fun buildOnBoardingService():OnBoardingService
 }
